+++ crates/sui-network/src/lib.rs
@@ -6,7 +6,6 @@ use std::time::Duration;
 
 pub mod api;
 pub mod discovery;
-pub mod metrics;
 pub mod randomness;
 pub mod state_sync;
 pub mod utils;
