+++ crates/sui-simulator/Cargo.toml
@@ -11,7 +11,7 @@ workspace = true
 
 [dependencies]
 move-package.workspace = true
-
+mysten-network.workspace = true
 sui-framework.workspace = true
 sui-move-build.workspace = true
 sui-types.workspace = true
