+++ crates/sui-simulator/src/lib.rs
@@ -16,6 +16,7 @@ pub use anemo_tower;
 pub use fastcrypto;
 pub use lru;
 pub use move_package;
+pub use mysten_network;
 pub use sui_framework;
 pub use sui_move_build;
 pub use sui_types;
